FEATURE: FET-390 Move package under NPM @gooddata org
The move under the organization requires package to be scoped by "@gooddata/" prefix, i.e., the package must be renamed. The name was chosen to match our tslint and eslint package names. Additionally, the license and readme file was updated to match the latest standards approved by our Legal department.